requirements.txt
aiohttp
asyncio
langchain-openai
python-dotenv
browser-use
aioredis
prometheus-client
pydantic